Enter test algorithm for calculating c and K
Just started to work with this for a small subset of the data. I used rapply to calculate the C values for all the priors, and then sketched out a cut and paste version to extract k for just one question. I need to make that more R-ish to get all values