Add code to calculate K from an input recursive list.
I wrote a function called calcK to input a recursive list, and then spit back out a flat list that just has the K values I'll use to sample from.

I also make a small, but important fix to the way the denominator was calculated. I'd been using the sum of the numerator, but that was one step past what I wanted from an input.