Add code to calculate the posterior using Jim Clark's algorithm from the Schick et al 2013 paper
Before when we'd been testing the algorithm, we had been using rdirichlet to just generate the draws. However, now we have a 9*9 matrix, and we will have changing input data (because of imputation) so we need to be able to sample using his algorithm.

This implements it for one question.