Substitute in the new priorList value in place of the old (incorrect) allqs.
The change I've made is to go from a 6 element recursive list in allqs, where each element was one of the 6 elicited questions, and each element was itself a recursive list with each expert's answers to a 2 element recursive list for males and females.

That way we've essentially squashed in all of the priors together as one slice.