Get mini-example working
What I've done here is to flesh out the sampling so that I show it works with a small example. I ran this with just 5 regions, instead of all 9.

The goal here was to use a series of nested loops to try and fill out the c matrix for all data/expert combinations.

I wrote the bulk of this last week, but have now gone back through to add some clarifying comments, and to make sure the code works.