Modify calcC() and add in transformC()
The goal of this was to take the exponentiation out of calcC() so that I didn't get any 0 values returned. Then I do the exponentiation in the new transformC() function. Before I take the exponenent, though, I check if any of the values in cmat are below a threshold. If so, I add a fudge factor to all of the experts in that region