Fix the vignette to account for Michail's comments 9/26/16.
Essentially the comments boiled down to the fact that when plotting the priors, I was using kmat to plot them, when I really wanted to be using 1 / numexp. I fixed that and updated the vignette to reflect that.

I deleted some incorrect text about the shape of the priors, and added in a clarifying comment about getting the data back from the sampler.